apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
Pillow==11.1.0
orjson==3.10.13
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
boto3==1.36.25
PyJWT==2.10.1
prometheus-client==0.21.1
